+++ newbee-mall-api-spring-boot-3.x/src/main/java/whut/pilipili/mall/api/mall/vo/PilipiliMallOrderItemVO.java
@@ -32,5 +32,5 @@ public class PilipiliMallOrderItemVO implements Serializable {
     private String goodsCoverImg;
 
     @Schema(title = "商品价格")
-    private Integer sellingPrice;
+    private Float sellingPrice;
 }
