+++ newbee-mall-api-spring-boot-3.x/src/main/java/whut/pilipili/mall/api/mall/vo/PilipiliMallOrderListVO.java
@@ -28,7 +28,7 @@ public class PilipiliMallOrderListVO implements Serializable {
     private String orderNo;
 
     @Schema(title = "订单价格")
-    private Integer totalPrice;
+    private Float totalPrice;
 
     @Schema(title = "订单支付方式")
     private Byte payType;
