+++ newbee-mall-api-spring-boot-3.x/src/main/java/whut/pilipili/mall/api/mall/vo/PilipiliMallSearchGoodsVO.java
@@ -32,6 +32,6 @@ public class PilipiliMallSearchGoodsVO implements Serializable {
     private String goodsCoverImg;
 
     @Schema(title = "商品价格")
-    private Integer sellingPrice;
+    private Float sellingPrice;
 
 }
