+++ newbee-mall-api-spring-boot-3.x/src/main/java/whut/pilipili/mall/api/mall/vo/PilipiliMallShoppingCartItemVO.java
@@ -35,5 +35,5 @@ public class PilipiliMallShoppingCartItemVO implements Serializable {
     private String goodsCoverImg;
 
     @Schema(title = "商品价格")
-    private Integer sellingPrice;
+    private Float sellingPrice;
 }
