+++ newbee-mall-api-spring-boot-3.x/src/main/java/whut/pilipili/mall/entity/PilipiliMallOrder.java
@@ -21,7 +21,7 @@ public class PilipiliMallOrder {
 
     private Long userId;
 
-    private Integer totalPrice;
+    private Float totalPrice;
 
     private Byte payStatus;
